update mempool reap amount

+++ mempool/clist_mempool.go
@@ -651,7 +651,7 @@ func (mem *CListMempool) ReapMaxBytesMaxGas(maxBytes, maxGas int64, fallbackMode
 
 	mem.metrics.GasReap.Set(float64(totalGas))
 	if mem.Size() > 0 {
-		mem.metrics.MempoolReapedPercent.Set(float64((len(txs) / mem.Size()) * 100))
+		mem.metrics.MempoolReapedPercent.Set((float64(len(txs)) / float64(mem.Size()) * 100))
 	} else {
 		mem.metrics.MempoolReapedPercent.Set(0.0)
 	}
